test: 관리자 권한 ReservationTime 조작 api 테스트 추가
